+++ server/v2/config.go
@@ -1,15 +1,27 @@
 package serverv2
 
-import "github.com/spf13/cobra"
+import (
+	"fmt"
 
-// CLIConfig defines the CLI configuration for a module server.
-type CLIConfig struct {
-	// Commands defines the main command of a module server.
-	Commands []*cobra.Command
-	// Queries defines the query commands of a module server.
-	// Those commands are meant to be added in the root query command.
-	Queries []*cobra.Command
-	// Txs defines the tx commands of a module server.
-	// Those commands are meant to be added in the root tx command.
-	Txs []*cobra.Command
+	"github.com/spf13/viper"
+)
+
+// ReadConfig returns a viper instance of the config file
+func ReadConfig(configPath string) (*viper.Viper, error) {
+	v := viper.New()
+	v.SetConfigType("toml")
+	v.SetConfigName("config")
+	v.AddConfigPath(configPath)
+	if err := v.ReadInConfig(); err != nil {
+		return nil, fmt.Errorf("failed to read config: %s: %w", configPath, err)
+	}
+
+	v.SetConfigName("app")
+	if err := v.MergeInConfig(); err != nil {
+		return nil, fmt.Errorf("failed to merge configuration: %w", err)
+	}
+
+	v.WatchConfig()
+
+	return v, nil
 }
